fix: App.tsx fix (#3)
- App.tsx에  있던 레이아웃, 네비바를 각 폴더 파일로 분리
