+++ package.json
@@ -16,6 +16,7 @@
     "@emotion/styled": "^11.13.0",
     "@tanstack/react-query": "^5.56.2",
     "@types/react-router-dom": "^5.3.3",
+    "lucide-react": "^0.441.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-router-dom": "^6.26.2"
