+++ src/Layouts/Navbar.tsx
@@ -1,0 +1,4 @@
+// 네비게이션바
+// 하단이 아닌 왼쪽 사이드에 위치해 있어야함.
+const Navbar = () => <div>네비게이션바</div>;
+export default Navbar;
